requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
transformers>=4.37.0
google-generativeai>=0.3.0
huggingface-hub>=0.20.0
unsloth  # Optional: 2x faster inference + 4-bit quantization for Qwen models

# Data Processing
pandas>=2.0.0
datasets>=4.0.0

# Search Tools (for Deep Research experiments)
duckduckgo-search>=4.0.0  # Free search, no API key needed
tavily-python  # Optional: Higher quality search, requires TAVILY_API_KEY

# Utilities
python-dotenv>=1.0.0

# Visualization (optional, for analyze_results.py)
matplotlib>=3.7.0
seaborn>=0.12.0

# Note: Install with:
# pip install -r requirements.txt
#
# For GPU support (CUDA):
# pip install torch --index-url https://download.pytorch.org/whl/cu118